Moved upload logic to UploadSection.tsx
